Rename some old components and remove unused ones
