+++ src/components/InstrumentControls/Camera.vue
@@ -171,7 +171,7 @@
 <script>
 import { commands_mixin } from '../../mixins/commands_mixin'
 import { user_mixin } from '../../mixins/user_mixin'
-import CommandButton from '@/components/CommandButton'
+import CommandButton from '@/components/FormElements/CommandButton'
 import StatusColumn from '@/components/status/StatusColumn'
 import StatusVal from '@/components/status/StatusVal'
 import SimpleDeviceStatus from '@/components/status/SimpleDeviceStatus'
