+++ src/components/calendar/UserEventsTable.vue
