+++ src/components/sitepages/SiteHome.vue
@@ -43,14 +43,12 @@
 
 
 <script>
-import TheDeviceSelectors from '@/components/TheDeviceSelectors'
-import TheDomeCam from '@/components/TheDomeCam'
 import { mapGetters, mapState } from 'vuex'
 import { commands_mixin } from '../../mixins/commands_mixin'
 import { user_mixin } from '../../mixins/user_mixin'
 import LeafletMap from '@/components/maps/LeafletMap'
 import SiteEventsModal from '@/components/SiteEventsModal'
-import CommandButton from '@/components/CommandButton'
+import CommandButton from '@/components/FormElements/CommandButton'
 
 export default {
     name: "SiteHome",
@@ -58,8 +56,6 @@ export default {
     mixins: [commands_mixin, user_mixin],
     components: {
         CommandButton, 
-        TheDeviceSelectors,
-        TheDomeCam,
         LeafletMap,
         SiteEventsModal,
     },
